==> Notebooks/1 - Webscraping the Hot 100 Billboard.ipynb <==
### Import libraries

from bs4 import BeautifulSoup
import requests
import pandas as pd

### Finding url and downloading and parsing html

url = "https://www.billboard.com/charts/hot-100"

response = requests.get(url)

soup = BeautifulSoup(response.content, "html.parser")

### Creating a list for each of the variables

# List for the variable song_title
song_title = []
for song in soup.select("span.chart-element__information__song.text--truncate.color--primary"):
    song_title.append(song.get_text())

#List for the variable artist
artist = []
for i in soup.select("span.chart-element__information__artist.text--truncate.color--secondary"):
    artist.append(i.get_text())

#List for last week ranking
last_week = []
for numb in soup.select("span.chart-element__meta.text--center.color--secondary.text--last"):
    last_week.append(numb.get_text())

# List for peak 
peak_list = []
for peak in soup.select("span.chart-element__meta.text--center.color--secondary.text--peak"):
    peak_list.append(peak.get_text())

# List for weeks on chart
weeks_on_chart = []
for week in soup.select("span.chart-element__meta.text--center.color--secondary.text--week"):
    weeks_on_chart.append(week.get_text())

### Creating the dataframe

songs = pd.DataFrame({
    "name":song_title, 
    "artist":artist,
    "last_week_ranking":last_week,
    "peak":peak_list,
    "weeks_on_chart":weeks_on_chart
    })

songs.to_csv("../Datasets/top_hot_100.csv", index=False)
