==> Notebooks/2 - Prototype 1.ipynb <==
### Importing libraries and reading files

import pandas as pd
import random

top_100_songs = pd.read_csv("../Datasets/top_hot_100.csv")

# Formatting the name and artist str
top_100_songs["name"] = top_100_songs["name"].apply(lambda x: x.lower())
top_100_songs["artist"] = top_100_songs["artist"].apply(lambda x: x.lower())

# Looping through the df to check if the song is hot or not

for song in top_100_songs[["name","artist"]].iterrows():
    user_input = input("Write down a song or artist to get a recommendation: ")
    if user_input.lower() in list(top_100_songs["name"]):
        random_song = random.choice(top_100_songs["name"]).title()
        song_artist = top_100_songs.loc[top_100_songs["name"] == random.choice(top_100_songs["name"]), "artist"].values[0]
        
        print("You may also like this song: ", random_song, "from", song_artist.title())
        print("Thanks for using the music recommender.")
        break
    elif user_input.lower() in list(top_100_songs["artist"]) or top_100_songs["artist"].str.contains(user_input).any():
        print("You may also like this artist: ", random.choice(top_100_songs["artist"]).title())
        print("Thanks for using the music recommender.")
        break
    elif user_input not in top_100_songs[["name","artist"]].iterrows():
        print("Sorry, we cannot make any recommendations of this song yet. Try with another hot song or artist:")
        continue
